Change nativeStepFrame return type to IntArray
Updated nativeStepFrame return type from ByteArray to IntArray.

+++ app/src/main/java/com/moonlight/moongba/EmuCore.kt
@@ -1,11 +1,11 @@
 package com.moonlight.moongba
 
 object EmuCore {
-    init { 
-        System.loadLibrary("gba_core") 
+    init {
+        System.loadLibrary("gba_core")
     }
 
     @JvmStatic external fun nativeInit(): Boolean
     @JvmStatic external fun nativeLoadRom(rom: ByteArray): Boolean
-    @JvmStatic external fun nativeStepFrame(): ByteArray
+    @JvmStatic external fun nativeStepFrame(): IntArray  // Changed from ByteArray
 }
